Only show factory reset on desktop for now
